deployment/requirements.txt
Flask==1.1.1
python-dotenv==0.10.3
redis==3.3.6
requests==2.21.0
